Move UserSession into lib
