--- auth/session/index.ts
@@ -1,2 +1,0 @@
-export * from "./AuthBanner"
-export * from "./UserSessionContext"
